Refactor packages
now each business logic package is expected to have its "handlers"
package which contains respective URL handlers
