+++ database/mysql.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package engine
+package database
 
 import (
 	"database/sql"
